feat: check synchronization of local dataset w/GBIF
Report the success or failure of a dataset creation or update operation
to alert users of synchronization issues. Define success and failure by
comparing the publication date of the local dataset EML metadata
and the endpoint of the zip archive download, with that of the remote
GBIF instance.

Move get_local_dataset_endpoint to utilities.py to prevent a circular
reference.