refactor: eliminate useless '_has_metadata' function
Remove the '_has_metadata' function as it does not serve a purpose.
Initially, it was designed to determine whether a local dataset group
had a member on GBIF and was used to guide decision logic concerning
resetting dataset endpoints and re-uploading metadata in the event of a
dataset update. However, it became apparent that this function returned
 'True' even if only boilerplate stand-in metadata was posted to GBIF
before the actual metadata was posted during a crawl operation.