ci: fix failing release workflow
Update the failing release workflow (GitHub Action) to enable Python
Semantic Release to automatically create a new release when the
development branch is merged into main.

- Add the GITHUB_TOKEN to the semantic release step, where it is
necessary for committing changes, using the standard syntax for
referencing secrets.
- Include the missing versioning command in the semantic release step to
ensure the new version is calculated.
- Correct the outdated syntax for referencing the version number in the
pyproject.toml file to align with the requirements of the current version
of Python Semantic Release.